+++ Lab3/Pages/Index.cshtml
@@ -10,7 +10,7 @@
     <div class="container mt-5">
         <div class="jumbotron">
             <h2 class="display-4">Brought to You By Tech Titans</h2>
-            <p class="lead">Jordon Fischer, Luke Evans, Divij Kumar, Dylan Kasher, Trevor Hudson, Gabe Oduro</p>
+            <p class="lead">Parker Adams, Luke Evans, Jordon Fischer, Alex Fish, Aidan Froehlich, Trevor Hudson, Dylan Kasher, Divij Kumar, Gabe Oduro, Alex Truong, Jeremy Wagner</p>
         </div>
     </div>
 </div>
